type modification fps

+++ src_base/xeve_rc.c
@@ -176,7 +176,7 @@ void xeve_init_rc_bpf_tbl(XEVE_CTX * ctx)
 
     int ld_struct = ctx->param.ref_pic_gap_length;
     int fnum_in_sec[10];
-    int ngop_in_sec = (ctx->param.fps + ld_struct - 1) / ld_struct;
+    int ngop_in_sec = ((int)ctx->param.fps + ld_struct - 1) / ld_struct;
 
     for (int i = ld_struct; i > 0; i = i >> 1)
     {
@@ -531,7 +531,7 @@ double get_qfactor_clip(XEVE_CTX *ctx, XEVE_RCORE * rcore, double qf)
         if (ctx->param.lookahead >= 24)
         {
             /* when encount scene change just after IDR, raise up qp to bits */
-            t0 = ctx->param.fps >> 3;
+            t0 = (int)ctx->param.fps >> 3;
             t1 = i_period >> 3;
 
             thd_distance = XEVE_MIN(t0, t1);
@@ -559,7 +559,7 @@ double get_qfactor_clip(XEVE_CTX *ctx, XEVE_RCORE * rcore, double qf)
         if (rcore->scene_type == SCENE_HIGH)
         {
             /* when encount scene change just before IDR, raise up qp to bits */
-            t0 = ctx->param.fps >> 3;
+            t0 = (int)ctx->param.fps >> 3;
             t1 = i_period >> 3;
 
             thd_distance = XEVE_MIN(t0, t1);
